requirements.txt
matplotlib>=3.1.3
torch>=1.10.0
pytorch_lightning>=1.4.9
torchmetrics>=0.5.1
kornia>=0.6.1
kornia_moons>=0.1.9
opencv-python>=4.5.4
numpy>=1.19.5
albumentations>=1.1.0
deepdish>=0.3.7
omegaconf>=2.1.1
shutup>=0.1.3
PyYAML>=5.4.1
wandb>=0.12.11